LPG_S/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
matplotlib>=3.4.0
PyYAML>=6.0
